Improve interface, cleanup
